Tidy up the image and actions packages
