+++ pkg/replacer/image/image_test.go
